rename: AppDelegate, SceneDelegate 를 Entry 폴더로 이동, 필요없는 주석 제거
